+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+* [ENHANCEMENT] Add support for the `PrometheusText1.0.0` scrape protocol. #7085
+* [BUGFIX] Add `goGC` field to `PrometheusAgent` CRD. #6667
+
 ## 0.78.2 / 2024-11-21
 
 * [BUGFIX] Revert the removal of the `proxyURL` field in `AlertmanagerConfig` CRD. The field takes precedence over `proxyUrl`. #7114